Move types to utils
