+++ app/src/components/market/common/list_item/index.tsx
@@ -6,7 +6,7 @@ import styled from 'styled-components'
 
 import { useGraphMarketMakerData } from '../../../../hooks'
 import { useConnectedWeb3Context } from '../../../../hooks/connectedWeb3'
-import { MarketMakerDataItem } from '../../../../queries/markets_home'
+import { MarketMakerDataItem } from '../../../../util/types'
 import { ERC20Service } from '../../../../services'
 import { calcPrice, formatBigNumber } from '../../../../util/tools'
 import { IconStar } from '../../../common/icons/IconStar'
