+++ app/src/queries/markets_home.ts
@@ -1,4 +1,3 @@
-import { BigNumber } from 'ethers/utils'
 import gql from 'graphql-tag'
 
 import { MarketFilters, MarketStates, MarketValidity, MarketsSortCriteria } from './../util/types'
@@ -18,34 +17,6 @@ export const MarketDataFragment = gql`
     scaledLiquidityParameter
   }
 `
-
-export type GraphMarketMakerDataItem = {
-  id: string
-  collateralVolume: string
-  collateralToken: string
-  outcomeTokenAmounts: string[]
-  title: string
-  outcomes: Maybe<string[]>
-  openingTimestamp: string
-  arbitrator: string
-  category: string
-  templateId: string
-  scaledLiquidityParameter: string
-}
-
-export type MarketMakerDataItem = {
-  address: string
-  collateralVolume: BigNumber
-  collateralToken: string
-  outcomeTokenAmounts: BigNumber[]
-  title: string
-  outcomes: Maybe<string[]>
-  openingTimestamp: Date
-  arbitrator: string
-  category: string
-  templateId: number
-  scaledLiquidityParameter: number
-}
 
 export const DEFAULT_OPTIONS = {
   state: MarketStates.open,
